+++ src/libraries/System.Private.CoreLib/src/System/DateTimeOffset.cs
@@ -228,7 +228,7 @@ namespace System
         public DateTimeOffset(int year, int month, int day, int hour, int minute, int second, int millisecond, int microsecond, TimeSpan offset)
             : this(year, month, day, hour, minute, second, millisecond, offset)
         {
-            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual((uint)microsecond, DateTime.MicrosecondsPerMillisecond);
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual((uint)microsecond, (uint)DateTime.MicrosecondsPerMillisecond);
             _dateTime = _dateTime.AddMicroseconds(microsecond);
         }
 
@@ -302,7 +302,7 @@ namespace System
         public DateTimeOffset(int year, int month, int day, int hour, int minute, int second, int millisecond, int microsecond, Calendar calendar, TimeSpan offset)
         : this(year, month, day, hour, minute, second, millisecond, calendar, offset)
         {
-            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual((uint)microsecond, DateTime.MicrosecondsPerMillisecond);
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual((uint)microsecond, (uint)DateTime.MicrosecondsPerMillisecond);
             _dateTime = _dateTime.AddMicroseconds(microsecond);
         }
