+++ src/libraries/System.Private.CoreLib/src/System/String.Manipulation.cs
@@ -506,7 +506,7 @@ namespace System
         {
             ArgumentNullException.ThrowIfNull(value);
 
-            ArgumentOutOfRangeException.ThrowIfGreaterThan((uint)startIndex, Length);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan((uint)startIndex, (uint)Length);
 
             int oldLength = Length;
             int insertLength = value.Length;
@@ -870,7 +870,7 @@ namespace System
         // a remove that just takes a startindex.
         public string Remove(int startIndex)
         {
-            ArgumentOutOfRangeException.ThrowIfGreaterThan((uint)startIndex, Length);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan((uint)startIndex, (uint)Length);
 
             return Substring(0, startIndex);
         }
